requirements.txt
streamlit==1.31.1
langchain==0.1.6
langchain-openai==0.0.5
langchain-google-genai==0.0.6
python-dotenv==1.0.1
Pillow==10.2.0
pytesseract==0.3.10
pdf2image==1.17.0
pypdf==4.0.1
openai==1.12.0
google-generativeai==0.3.2
pydantic==2.6.1
